Add first implementation of srp/2
Solves the stable roommates problem, at least on a couple of examples.
Some examples are provided as test cases.
This implementation seems to be too slow for larger instances. 10 people
example takes more than a minute on my computer. Note that I aborted this
long run before it got to terminate.